Fixed driver change bug, added proper While loop
So it turns out I was feeding the wrong information to the SDK when doing driver camera changes. I was feeding the index number of the driver's name, not the car number. iRacing wants the car number. Fixed.

Also added a While statement to sleep the app for 30 seconds if iRacing is not running. My earlier dumb if statement would obviously have failed after 30 seconds.