More reworking & cleanup
Rearranged the columns a bit at the bottom to make a bit more sense. Re-added the change position menu, might move that to buttons in the future, but this layout is already looking pretty cramped. Added some global variables to some of the functions to make things look cleaner, also used a much easier "weekendinfo" key to get team racing information.